fix: PageSection enum moved to shared types to resolve circular dependency issue
