+++ .gitignore
@@ -23,6 +23,7 @@ dist-ssr
 *.sln
 *.sw?
 
+# TypeScript build info
 tsconfig.tsbuildinfo
 
 # Firebase
