+++ src/types/pageSection.ts
@@ -1,0 +1,7 @@
+export enum PageSection {
+    TRIPS = 'trips',
+    FAVORITES = 'favorites',
+    PROFILE = 'profile',
+    PAYMENT_MANAGE = 'paymentMethods',
+    COUPONS = 'coupons',
+}
