--- tsconfig.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/atoms/auth.ts","./src/atoms/booking.ts","./src/atoms/cart.ts","./src/atoms/countrywithcities.ts","./src/atoms/modal.ts","./src/atoms/search.ts","./src/atoms/tours.ts","./src/atoms/trendingtours.ts","./src/atoms/wishlist.ts","./src/components/copylink.tsx","./src/components/filterbar.tsx","./src/components/footer.tsx","./src/components/header.tsx","./src/components/notfound.tsx","./src/components/partnerform.tsx","./src/components/scolltotop.tsx","./src/components/searchbar.tsx","./src/components/steps.tsx","./src/components/datatypes.tsx","./src/components/card/bannercard.tsx","./src/components/card/tripcard.tsx","./src/components/card/type.ts","./src/components/cart/enterinfostep.tsx","./src/components/map/mapmarker.tsx","./src/components/map/mappopup.tsx","./src/components/modal/modal.tsx","./src/components/modal/modalprovider.tsx","./src/components/modal/userprofilemodal.tsx","./src/components/modal/index.tsx","./src/components/modal/headermodals/cartmodal.tsx","./src/components/modal/headermodals/countriesmodal.tsx","./src/components/modal/headermodals/currencymodal.tsx","./src/components/modal/headermodals/languagemodal.tsx","./src/components/modalpopup/modalalert.tsx","./src/components/modalpopup/index.tsx","./src/components/modalpopup/authmodal/modalauth.tsx","./src/components/modalpopup/authmodal/passwordcomponent.tsx","./src/components/modalpopup/authmodal/phoneverification.tsx","./src/components/modalpopup/authmodal/sociallogin.tsx","./src/components/modalpopup/authmodal/telegramotpverify.tsx","./src/components/modalpopup/coremodal/coremodalpopup.tsx","./src/components/password/forgotpassword.tsx","./src/components/payment/bookingconfirmation.tsx","./src/components/payment/guestform.tsx","./src/components/payment/ordersummary.tsx","./src/components/payment/paymentreturnhandle.tsx","./src/components/payment/paymentstep.tsx","./src/components/payment/stripecardform.tsx","./src/components/common/atomicdropdownelements.tsx","./src/components/common/button.tsx","./src/components/common/card.tsx","./src/components/common/dropdownelemstyles.ts","./src/components/common/input.tsx","./src/components/common/nodatafound.tsx","./src/components/common/select.tsx","./src/components/common/skeletonloader.tsx","./src/components/common/tabs.tsx","./src/components/destinations/destinationcard.tsx","./src/components/home/firsthome.tsx","./src/components/home/popularaccomadations.tsx","./src/components/home/populardestinations.tsx","./src/components/home/recentlyaddedtours.tsx","./src/components/home/trendingtours.tsx","./src/components/layout/layout.tsx","./src/components/tours/tourstatustabs.tsx","./src/components/tours/tourscard.tsx","./src/constants/countrycodes.ts","./src/constants/countrywithcities.ts","./src/constants/statusmessages.ts","./src/context/authprovider.tsx","./src/data/mockdata.ts","./src/hooks/api/usedestinationcache.ts","./src/hooks/api/usefetch.ts","./src/hooks/api/useprofilecache.ts","./src/hooks/api/usestatushandler.ts","./src/hooks/api/usetrendingtours.ts","./src/hooks/api/usetripactions.ts","./src/hooks/api/usetypedfetch.ts","./src/hooks/queries/userprofilequery.ts","./src/hooks/ui/usedropdownmodal.ts","./src/hooks/ui/usefavorite.ts","./src/hooks/ui/usemodal.ts","./src/hooks/ui/usepaypalreturn.ts","./src/hooks/utils/usesearchactions.ts","./src/hooks/utils/usevalidation.ts","./src/pages/oauthredirect.tsx","./src/pages/cart/cart.tsx","./src/pages/checkout/cancel/paymentcancel.tsx","./src/pages/checkout/return/paymentreturn.tsx","./src/pages/contact/contactus.tsx","./src/pages/home/cityexpert.tsx","./src/pages/home/newsletter.tsx","./src/pages/home/index.tsx","./src/pages/home/destinations/topdestinationsbycity.tsx","./src/pages/home/destinations/topdestinationslist.tsx","./src/pages/legal/cancellationpolicy.tsx","./src/pages/legal/legalpageslayout.tsx","./src/pages/legal/privacypolicy.tsx","./src/pages/legal/termsconditions.tsx","./src/pages/mypage/myfavorites.tsx","./src/pages/mypage/mytrip.tsx","./src/pages/mypage/profilecontent.tsx","./src/pages/mypage/sidebar.tsx","./src/pages/mypage/index.tsx","./src/pages/mypage/order/orderhistory.tsx","./src/pages/partner/becomepartner.tsx","./src/pages/searchpackage/mapbox.tsx","./src/pages/searchpackage/searchpackagecontent.tsx","./src/pages/searchpackage/searchpackagedetails.tsx","./src/pages/searchpackage/searchpackagelist.tsx","./src/pages/searchpackage/searchpackagesidebar.tsx","./src/pages/tours/trendingtours.tsx","./src/providers/queryproviders.tsx","./src/routes/protectedroute.tsx","./src/routes/publicroutes.tsx","./src/routes/routes.tsx","./src/services/stripe.ts","./src/services/api/index.ts","./src/services/api/auth/authservice.ts","./src/services/api/auth/index.ts","./src/services/api/auth/types.ts","./src/services/api/cart/index.ts","./src/services/api/cart/types.ts","./src/services/api/cart/usecartservice.ts","./src/services/api/checkout/checkoutservice.ts","./src/services/api/checkout/index.ts","./src/services/api/checkout/types.ts","./src/services/api/client/httpclient.ts","./src/services/api/client/index.ts","./src/services/api/tours/index.ts","./src/services/api/tours/toursservice.ts","./src/services/api/tours/types.ts","./src/services/api/user/index.ts","./src/services/api/user/types.ts","./src/services/api/user/userservice.ts","./src/services/api/wishlist/index.ts","./src/services/api/wishlist/types.ts","./src/services/api/wishlist/wishlistservice.ts","./src/services/cache/cookieauthservice.tsx","./src/styles/globalstyles.ts","./src/styles/declaretheme.ts","./src/styles/theme.ts","./src/types/index.ts","./src/types/common/fetch.ts","./src/types/common/maxbox.ts","./src/types/common/modal.ts","./src/types/common/search.ts","./src/utils/geocodingcheck.ts","./src/utils/geocodingservice.ts","./src/utils/general/cartitemhandling.ts","./src/utils/general/formatcurrency.ts","./src/utils/general/tourchangetypes.ts","./src/utils/general/tourdetailshelper.ts","./src/utils/payment/paypalhelpers.ts","./src/utils/payment/paypalreturn.ts","./src/utils/tokenmanagement/cookieutils.ts","./src/utils/tokenmanagement/profilecachemanager.ts","./src/utils/tokenmanagement/tokenstorage.ts","./vite.config.ts"],"version":"5.7.3"}
